Miscellaneous warning cleanup
Assorted warning cleanup, that it didn't seem to make sense having one
patch per file for.

This patch
    - Adds some casting for syscall pointers
    - ANSIfies SRXAFSCB_GetDE in fsprobe
    - Loads sys/ioctl.h on some more platforms in usd
    - Includes some missing header files
    - Removes unused variables
    - Makes it clear that VIsSalvager does handle all enumerated types
    - Adds some more prototypes

Reviewed-on: http://gerrit.openafs.org/508
Tested-by: Derrick Brashear <shadow@dementia.org>

+++ src/cmd/cmd.c
@@ -232,7 +232,7 @@ PrintFlagHelp(register struct cmd_syndesc *as)
 {
     register int i;
     register struct cmd_parmdesc *tp;
-    size_t flag_width;
+    int flag_width;
     char *flag_prefix;
 
     /* find flag name length */
